+++ R/stratallo-package.R
@@ -1,0 +1,29 @@
+#' Optimal Sample Allocation in Stratified Sampling Schemes
+#'
+#' @name stratallo-package
+#' @aliases stratallo
+#' @docType package
+#' @title Functions for Optimal Sample Allocation in Stratified Sampling Schemes
+#' @author Wojciech Wojciak \email{wojciech.wojciak@@gmail.com}
+#' @import checkmate
+#' @importFrom lifecycle badge
+#' @keywords package
+#' @references
+#'   Wesołowski, J., Wieczorkowski, R., Wójciak, W. (2021),
+#'   Optimality of the recursive Neyman allocation,
+#'   *Journal of Survey Statistics and Methodology*,
+#'   \doi{10.1093/jssam/smab018},
+#'   \doi{10.48550/arXiv.2105.14486} \cr
+#'
+#'   Wójciak, W. (2022),
+#'   Minimum sample size allocation in stratified sampling under constraints on
+#'   variance and strata sample sizes,
+#'   \doi{10.48550/arXiv.2204.04035} \cr
+#'
+#'   Wójciak, W. (2019), Optimal allocation in stratified sampling schemes,
+#'   *MSc Thesis*, Warsaw University of Technology, Warsaw, Poland.
+#'   <http://home.elka.pw.edu.pl/~wwojciak/msc_optimal_allocation.pdf> \cr
+#'
+#'   Sarndal, C.-E., Swensson, B., and Wretman, J. (1992),
+#'   *Model Assisted Survey Sampling*, New York, NY: Springer.
+NULL