+++ temp_r_files/testy.R
@@ -1,0 +1,60 @@
+#################
+# dane
+#################
+
+# zbior data_mM jest w katalogu ./data
+# load(file = "./data/data_mM.Rdata")
+
+d <- data_mM[, "N"] * data_mM[, "S"]
+m <- data_mM[, "m"]
+M <- data_mM[, "M"]
+n_feasible <- (sum(m)+1):(sum(M)-1) # wektor feasible n (z wylaczeniem dwoch trywialnych przypadkow)
+
+#################
+# poprawnosc
+#################
+
+## 1)
+
+# n - skalar
+n <- sample(n_feasible, size = 1)
+noptcond_sufficient(d = d, l = m, u = M, n = n)
+rNa2_mM(d = d, m = m, M = M, n = n)
+
+# w funcji n, ktore przebiega przez wszystkie elementy w wektorze n_feasible
+nopt_ <- sapply(n_feasible, function(n_i) {
+  all.equal(
+    noptcond_sufficient(d = d, l = m, u = M, n = n_i),
+    rNa2_mM(d = d, m = m, M = M, n = n_i)
+  )
+}
+)
+all(nopt_)
+
+## 2)
+
+m <- sample(x = 10:10^3, size = 10^3, replace = TRUE)
+M <- m + sample(x = 10:10^2, size = length(m), replace = TRUE)
+d <- sample(x = 10:10^2, size = length(m), replace = TRUE)
+n_feasible <- (sum(m)+1):(sum(M)-1) # wektor feasible n
+
+nopt_ <- sapply(n_feasible, function(n_i)
+  all.equal(
+    noptcond_sufficient(d = d, l = m, u = M, n = n_i),
+    rNa_mM(d = d, m = m, M = M, n = n_i)
+  )
+)
+all(nopt_)
+
+#################
+# czasy obliczen
+#################
+
+# n - skalar
+n <- sample(n_feasible, size = 1)
+bench::mark(
+  noptcond_sufficient(d = d, l = m, u = M, n = n),
+  rNa_mM(d = d, m = m, M = M, n = n),
+  check = TRUE,
+  iterations = 10
+)[1:6]