+++ NEWS.md
@@ -1,5 +1,14 @@
 # stratallo
 
+# stratallo 2.0.0
+
+* Added new user function `nopt` for minimization of total sample size under
+  constraints on variance and upper bounds on strata sample sizes. 
+* Re-factored `rNa` into `rna_one_sided` so that it also computes the allocation
+  under optional lower bounds constraints.
+* Renamed `nopt` do `dopt` to reflect the fact that it minimizes variance
+  function D.
+  
 # stratallo 0.1.0
 
 * First release.